% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Create forest plot from meta analysis results}
\usage{
forest_plot(meta_analysis_results, effect_size = "rucsios_A",
  baseline_trend_exclusion_criterion = NULL)
}
\arguments{
\item{baseline_trend_exclusion_criterion}{If set to a numeric value, cases with a baseline trend (ie standardized beta OLS regression value for the timepoint A scores) with an absolute value greater than this numeric value will be excluded from the meta analysis.}

\item{meta_analysis}{Output of sced_meta_analysis().}
}
\description{
Meta analyse the standardized effect sizes (Ruscio's A or Hedges' g) provided by sced_analyse.
}
\examples{
# meta analysis
meta_analysis <- sced_meta_analysis(results = sced_results, effect_size = "ruscios_A")

# forest plot
forest_plot(meta_analysis)

}
