% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pval_better.R
\name{format_pval_better}
\alias{format_pval_better}
\title{Format p values better for printing and reporting}
\usage{
format_pval_better(p)
}
\arguments{
\item{value}{The p value to be formatted}
}
\description{
Tries to do a better job that base R's format.pval
}
\examples{
format_pval_better(0.00000001)
format_pval_better(0.01)
format_pval_better(0.7)
}
