% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity_metrics_for_forest.r
\name{heterogeneity_metrics_for_forest}
\alias{heterogeneity_metrics_for_forest}
\title{heterogeneity_metrics_for_forest}
\usage{
heterogeneity_metrics_for_forest(fit)
}
\arguments{
\item{fit}{The meta analysis fit object outputted by sced_meta_analysis()}
}
\description{
This function converts a metafor metaanalysis fit object to a string that contains the I^2 and H^2 heterogeneity statistics so that it can easily be added to a metafor forest plot.
}
\examples{
# calculate participant level results
results <- sced_analysis(data = data)

# fit meta
sced_meta_fit <- sced_meta_analysis(results = results, effect_size = "ruscios_A")

# forest plot
metafor::forest(sced_meta_fit$model_fit,
                xlab = "Probability of superiority",
                transf = boot::inv.logit,  # convert logits back to probabilities (ie Ruscio's A)
                mlab = heterogeneity_metrics_for_forest(sced_meta_fit$model_fit),
                digits = 2,
                addcred = TRUE,
                refline = 0.5)
                
}
