% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logodds_to_probability.r
\name{logodds_to_probability}
\alias{logodds_to_probability}
\title{logodds_to_probability}
\usage{
logodds_to_probability(x)
}
\arguments{
\item{x}{The integer(s) to be converted}
}
\description{
This function converts log odds probability values.
}
\examples{
logodds_to_probability()
}
