% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_meta.R
\name{power_meta}
\alias{power_meta}
\title{A priori power analysis for meta analysis}
\usage{
power_meta(yi, ni, k, tau2)
}
\arguments{
\item{yi}{Meta-analyzed hedges' g effect size. NBs 1) Although Cohen's rules of thumb (.2 = small, .5 = medium, .8 = large), larger effect sizes are often observed in SCED due to the within subject designs (e.g., above 1.0). 2) Valentine's equations assume a between-groups design but here we apply to within-subjects effect sizes. Power is therefore likely to be underestimated. 3) Valentine's equations employ Cohen's d effect size but we use the Hedges' g effect size which has been bias corrected for small samples.}

\item{ni}{The average number of data points per phase (i.e., in each of A and B, before or after the intervention). Assuming equal number of total time points for all participants, this is roughly equal to half the total number of time points. NB again, Valentine's equations assume a between-groups design but here we apply to within-subjects effect sizes. Power is therefore likely to be underestimated.}

\item{k}{The number of participants.}

\item{tau2}{The tau^2 metric of between participant heterogeneity, as produced by the meta analysis function. If data from previous studies is not available, historically reccommended cutoffs are .33 for small, 1.0 for moderate, & 3.0 for large.}
}
\value{
Statistical power (i.e., 1 - Beta), the probability of observing a significant result given the parameters.
}
\description{
Rough approximation of power for the between participant meta analysis of effects, using equations derived from Valentine, Pigott, & Rothstein (2010, doi: 10.3102/1076998609346961), derived by Quintana (2017: https://towardsdatascience.com/how-to-calculate-statistical-power-for-your-meta-analysis-e108ee586ae8), and applied to within-subject SCED designs here.
}
\examples{
power_meta(yi = 0.5, 
           ni = 25,
           k  = 6,
           tau2 = 1)
}
