% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_df.r
\name{round_df}
\alias{round_df}
\title{round_df}
\usage{
round_df(df, digits)
}
\arguments{
\item{df}{Dataframe to be rounded.}

\item{digits}{Number of decimal places to round to.}
}
\description{
This function rounds all numeric variables in a dataframe to the desired number of places. Non-numeric variables will be ignored.
}
\examples{
round_df(my_data, digits = 2)
}
