% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruscios_A.R
\name{ruscios_A}
\alias{ruscios_A}
\title{Ruscio's A - the probability of superiority}
\usage{
ruscios_A(
  data,
  variable,
  group,
  value1 = 1,
  value2 = 0,
  adjust_ceiling = FALSE
)
}
\arguments{
\item{data}{data}

\item{variable}{continuous variable}

\item{group}{dichotomous group}

\item{value1}{assignment of group 1}

\item{value2}{assignment of group 2}
}
\value{
Ruscio's A.
}
\description{
This function bootstraps confidence intervals for Ruscio's A effect size (2008). 
Code adapted from adapted from John Ruscio's original implementation of his metric: https://ruscio.pages.tcnj.edu/quantitative-methods-program-code/
}
\examples{
ruscios_A(data = simulated_data, variable = "Score", group = "Condition", value1 = "B", value2 = "A")

}
