% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruscios_A_boot.R
\name{ruscios_A_boot}
\alias{ruscios_A_boot}
\title{Bootstrapped Ruscio's A}
\usage{
ruscios_A_boot(data, variable, group, value1, value2, B = 1000)
}
\arguments{
\item{data}{data}

\item{variable}{continuous variable}

\item{group}{dichotomous group}

\item{value1}{assignement of group 1}

\item{value2}{assignement of group 2}

\item{B}{Number of boostrapped resamples}
}
\value{
ruscios_A_median Median boostrapped estimation of Ruscio's A.

ruscios_A_ci_lwr Lower 95% boostrapped confidence interval

ruscios_A_ci_upr Upper 95% boostrapped confidence interval
}
\description{
This function bootstraps confidence intervals for Ruscio's A effect size (2008).
}
\examples{
ruscios_A_boot(variable = "Score", group = "Condition", value1 = "B", value2 = "A", B = 1000, data = simulated_data)
}
