% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCED.r
\docType{package}
\name{SCED-package}
\alias{SCED}
\alias{SCED-package}
\title{Analyse and plot data from AB Single Case Design Experiments}
\description{
Analyses are conduced using non-parametric permutation tests and effect sizes
}
\details{
Use \code{\link{sced_plot}} to plot data. Use \code{\link{sced_analyse}} to analyse it.
The formatting of the data frame (including variable names and condition designators) that
must be passed to each of the above can be found in the included dataset \code{\link{simulated_data}}. Run View(simulated_data) to see this.
}
\author{
\strong{Maintainer}: Ian Hussey \email{ian.hussey@ugent.be}

}
