% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sced_analysis.r
\name{sced_analysis}
\alias{sced_analysis}
\title{Analyse data}
\usage{
sced_analysis(data)
}
\arguments{
\item{data}{Experiment data}
}
\description{
Analyse data from an AB design SCED experiment using non-parametric frequentist tests
}
\examples{
sced_analysis(simulated_data)
}
\keyword{SCED}
