% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sced_plot.r
\name{sced_plot}
\alias{sced_plot}
\title{Plot data}
\usage{
sced_plot(data)
}
\arguments{
\item{data}{Experiment data. This must contain columns named "Participant", "Timepoint" (integer), "Score" (numeric; your DV), and "Condition" (must include only "A" and "B" as a string or factor). See the included simulated_data dataset for an example using \code{View(simulated_data)}.}
}
\description{
Plot data from an AB design SCED experiment
}
\examples{
sced_plot(data = simulated_data)
}
