% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sced_summary.R
\name{sced_summary}
\alias{sced_summary}
\title{Summarise sced_analysis output in a table}
\usage{
sced_summary(results)
}
\arguments{
\item{results}{Output of \code{View(sced_analysis)}.}
}
\value{
a data frame with formatted results
}
\description{
Summarise output of sced_analysis in a table for printing.
}
\examples{
sced_summary(sced_results)
}
