% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_ab_data.r
\name{simulate_ab_data}
\alias{simulate_ab_data}
\title{Simulate AB SCED data}
\usage{
simulate_ab_data(participants, timepoints, parameters)
}
\arguments{
\item{participants}{number of participants to simulate}

\item{timepoints}{number of timepoints to simulate for both the A and B conditions}

\item{parameters}{parameters of the data to be simulated (e.g., true effect size)}
}
\description{
Simulate data for an AB SCED design using simstudy
An arbitrary number of timepoints can be simulated, and an arbitrary true effect size. An equal number of time points are created for the A and B conditions.
}
\examples{
# True effect size of Cohen's *d* = 1.5 and *A* = 0.85.
parameters_d_1.5 <-
  defData(varname = "A", dist = "normal", formula = 0,   variance = 1, id = "idnum") \%>\%
  defData(varname = "B", dist = "normal", formula = 1.5, variance = 1, id = "idnum")

sim <- simulate_ab_data(participants = 1000, timepoints = 15, parameters = parameters_d_1.5)
}
