% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.r
\name{simulate_data}
\alias{simulate_data}
\title{Simulate AB SCED data}
\usage{
simulate_data(participants = 10, timepoints = 15, cohens_d = 1.5)
}
\arguments{
\item{participants}{number of participants to simulate}

\item{timepoints}{number of timepoints to simulate for both the A and B conditions}

\item{cohens_d}{the true effect size to simulate. Default of Cohen's d = 1.5, which corrisponds to a Ruscio's A of 0.85. NB this is stated in Cohen's d rather than Hedges' g because g is sample size dependent where d is not.}
}
\description{
Simulate data for an AB SCED design using simstudy
An arbitrary number of timepoints can be simulated, and an arbitrary true effect size. An equal number of time points are created for the A and B conditions.
}
\examples{
# True effect size of Cohen's d = 1.5 / A = 0.85.
# simulate data using these parameters
sim <- simulate_ab_data(participants = 10, timepoints = 15, cohens_d = 1.5)
}
