% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.r
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated AB SCED data}
\format{A data frame with 300 rows and 5 variables, i.e., 15 data points in both A and B conditions and 10 participants.}
\source{
Generated using the simstudy R package.
}
\usage{
simulated_data
}
\description{
A simulated dataset sampled from a true distribution of Cohen's *d* = 1.5.
}
\keyword{datasets}
