% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{sced_meta_analysis}
\alias{sced_meta_analysis}
\title{Meta analyse probability values (Ruscio's A)}
\usage{
sced_meta_analysis(results)
}
\arguments{
\item{results}{Output of sced_analysis().}
}
\description{
Meta analyse the robust effect sizes (Ruscio's A) provided by sced_analyse
}
\examples{
# meta analysis
meta_analysis <- sced_meta_analysis(results = results)

# return results
meta_analysis$meta_effect
meta_analysis$meta_heterogeneity

# forest plot
forest(meta_analysis$model_fit, 
       xlab = "Odds ratio", 
       addcred = TRUE)  # credibility intervals
## add column headings to the plot
text(-6.3, 6.8, "Participant", pos = 4)
text( 8.4, 6.8, "Odds Ratio [95\% CI]", pos = 2)
}
