% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{sced_meta_analysis}
\alias{sced_meta_analysis}
\title{Meta analyse probability values (Ruscio's A)}
\usage{
sced_meta_analysis(results)
}
\arguments{
\item{results}{Output of sced_analyse.}
}
\description{
Meta analyse the robust effect sizes (Ruscio's A) provided by sced_analyse
}
\examples{
sced_meta_analysis(results = sced_results)
}
