% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sced_plot.r
\name{sced_plot}
\alias{sced_plot}
\title{Plot data}
\usage{
sced_plot(data)
}
\arguments{
\item{data}{Experiment data}
}
\description{
Plot data from an AB design SCED experiment
}
\examples{
sced_plot(simulated_data)
}
\keyword{SCED}
