% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.r
\docType{data}
\name{simulated_data}
\alias{simulated_data}
\title{Simulated AB SCED data}
\format{
A data frame with 180 rows and 4 variables, representing 6 participants with a verying number of timepoints in the A vs B conditions.
}
\source{
Generated using the simstudy R package.
}
\usage{
simulated_data
}
\description{
A simulated dataset sampled from a true distribution of Cohen's d = 1.5.
}
\keyword{datasets}
